requirements.txt
streamlit
langchain
langchain_community
langchain_core
langchain_groq
langchain-cohere
bs4
pypdf
faiss-cpu
PyPDF2
arxiv
python-dotenv